new data HandlersCache
